requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
pydantic
passlib[bcrypt]
pyjwt
pytest
httpx
ruff
